feat: move to dynamic import for the react-flow client
